64-битные смещения и передача причины отказа доставки в 1С
Смещение Kafka сужалось до int32 на границе с 1С: ПолучитьСмещениеСообщения
возвращало (int32_t) offset, а ЗафиксироватьСмещение, ПрочитатьСообщениеПоСмещению
и УстановитьПозициюЧтения читали параметр через std::get<int32_t>. На топиках со
смещением выше 2^31-1 это давало неверные значения и молчаливую повторную обработку
или пропуск сообщений. Добавлен общий конвертер variantToInt64: принимает целое и
VTYPE_R8 (1С маршалит значения свыше int32 как число с плавающей точкой), отвергает
дробные значения, NaN, бесконечность и величины свыше 2^53 с внятным текстом ошибки
вместо bad_variant_access.

Причина неуспешной синхронной отправки не доходила до 1С: отчёт о доставке писал
"Delivery failed: <errstr>" в слот асинхронных ошибок, а produceAndWaitResult
затирал его заглушкой "Not delivered. See log for details". Текст отчёта дополнен
кодом ошибки, топиком и партицией; заглушка больше не перекрывает его; слот
очищается перед каждой отправкой с ожиданием результата.

Спецификации обеих доработок добавлены в specs/.

+++ src/SimpleKafka1C.h
@@ -208,7 +208,10 @@ private:
 	variant_t getMessageData(const variant_t &binaryResult);	// данные, как они есть в kafka
 	std::string getMessageKey();
 	std::string getMessageHeaders();
-	int32_t getMessageOffset();
+	// Returned as double: variant_t has no int64 alternative, and 1C receives
+	// VTYPE_R8 as a plain Number. Offsets are int64 in Kafka, so narrowing to
+	// int32 corrupted values above 2^31.
+	double getMessageOffset();
 	std::string getMessageTopicName();
 	int32_t getMessageBrokerID();
 	double getMessageTimestamp();
